completed search to use universally on the web
