+++ client/pages/cart.jsx
@@ -17,12 +17,6 @@ import Breadcrumbs from '@mui/material/Breadcrumbs';
 import Link from '@mui/material/Link';
 import Stack from '@mui/material/Stack';
 import NavigateNextIcon from '@mui/icons-material/NavigateNext';
-import ImageListItem from '@mui/material/ImageListItem';
-import { FormControl, Input, InputAdornment } from '@mui/material';
-import SearchIcon from '@mui/icons-material/SearchOutlined';
-import Box from '@mui/material/Box';
-import Drawer from '@mui/material/Drawer';
-import CloseIcon from '@mui/icons-material/Close';
 
 const theme = createTheme({
   palette: {
@@ -90,25 +84,16 @@ export default class Cart extends React.Component {
     super(props);
     this.state = {
       cartItems: [],
-      products: [],
-      filteredProductsList: [],
       isOpen: false
     };
     this.orderSummary = this.orderSummary.bind(this);
     this.delete = this.delete.bind(this);
     this.CustomSeparator = this.CustomSeparator.bind(this);
-    this.onSearchInputChange = this.onSearchInputChange.bind(this);
-    this.filterProducts = this.filterProducts.bind(this);
     this.openModal = this.openModal.bind(this);
     this.closeModal = this.closeModal.bind(this);
-    this.searchModal = this.searchModal.bind(this);
   }
 
   componentDidMount() {
-    fetch('/api/shoes')
-      .then(res => res.json())
-      .then(products => this.setState({ products, filteredProductsList: products }))
-      .catch(err => console.error(err));
     const token = window.localStorage.getItem('token');
     if (token) {
       fetch('/api/cart', {
@@ -141,81 +126,6 @@ export default class Cart extends React.Component {
 
   closeModal() {
     this.setState({ isOpen: false });
-  }
-
-  onSearchInputChange(event) {
-    this.setState({ searchInput: event.target.value.toLowerCase() }, () => {
-      if (this.state.searchInput === '') {
-        this.setState({ filteredProductsList: this.state.products });
-      }
-    });
-  }
-
-  filterProducts(event) {
-    if (event.key === 'Enter' || event.type === 'click') {
-      let filteredArr = this.state.products.filter(product => product.name.toLowerCase().includes(this.state.searchInput.toLowerCase()));
-      if (this.state.searchInput === '') {
-        filteredArr = this.state.products;
-      }
-      this.setState({ filteredProductsList: filteredArr });
-      this.closeModal();
-    }
-  }
-
-  searchModal() {
-    const productList = this.state.filteredProductsList;
-    return (
-      <Drawer
-        {...this}
-        anchor='right'
-        open={this.state.isOpen}
-        onClose={this.closeModal}
-      >
-        <Box style={{ width: '390px' }}>
-          <span style={styles.xIcon}>
-            <CloseIcon onClick={this.closeModal} className='xIcon' />
-          </span>
-          <Container style={{ marginLeft: '1rem', justifyContent: 'center' }}>
-            <FormControl variant="standard" sx={{ m: 1, mt: 2, width: '300px' }} >
-              <Input placeholder='Search our store'
-                id="standard-adornment-weight"
-                endAdornment={<InputAdornment position="end"><SearchIcon onClick={this.filterProducts} style={{ cursor: 'pointer' }} /></InputAdornment>}
-                aria-describedby="standard-weight-helper-text"
-                inputProps={{
-                  'aria-label': 'weight',
-                  type: 'search'
-                }}
-                style={{ fontFamily: 'eczar' }}
-                onChange={this.onSearchInputChange}
-                onKeyDown={this.filterProducts}
-              />
-              <div style={{ lineHeight: '1rem' }}>
-                <p style={{ fontWeight: 'bold' }}>Popular searches</p>
-                <div style={{ marginTop: 0, paddingLeft: 0, display: 'flex' }}>
-                  <p className='brands'>Jordan Nike Yeezy New Balance</p>
-                </div>
-              </div>
-            </FormControl>
-          </Container>
-          <Grid style={{ marginTop: 0, marginLeft: '2rem' }}>
-            <ImageList style={{ marginTop: 0, width: '335px' }} >
-              {productList.map(item => (
-                <a href={`#product?product=${item.productId}`} style={{ textDecoration: 'none', color: 'black', width: '150px' }} key={item.productId} >
-                  <ImageListItem style={{ width: '150px' }}>
-                    <img style={{ width: '150px', height: '150px' }}
-                      src={item.imageUrl}
-                      srcSet={item.imageUrl}
-                      alt={item.title}
-                    />
-                  </ImageListItem>
-                </a>
-              ))}
-            </ImageList>
-          </Grid>
-
-        </Box>
-      </Drawer>
-    );
   }
 
   CustomSeparator() {
@@ -353,7 +263,6 @@ export default class Cart extends React.Component {
               </Paper>
             </Grid>
           </Container>
-          {this.searchModal()};
         </>
       );
     } else {
@@ -425,7 +334,6 @@ export default class Cart extends React.Component {
               </Grid>
             </Grid>
           </Container>
-          {this.searchModal()}
         </>
       );
     }
